functions/extract-rss/requirements.txt
functions-framework==3.*
google-cloud-storage
google-cloud-secret-manager
feedparser
requests==2.28.1